% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facettize.R
\name{facettize}
\alias{facettize}
\title{Flexible plot facetting}
\usage{
facettize(
  p,
  rows = NULL,
  cols = NULL,
  wrap = FALSE,
  prepend = NULL,
  append = NULL,
  parsed = NULL,
  header = NULL,
  sep = " = "
)
}
\arguments{
\item{p}{A ggplot}

\item{rows}{Optional variables to facet by rows}

\item{cols}{Optional variables to facet by columns}

\item{wrap}{If TRUE, \code{facet_wrap} is used instead of \code{facet_grid}, so columns and rows are irrelevant}

\item{prepend, append}{Optional character strings to prepend or append to the labels, respectively. Can be either of (1) an unnamed character string that will be recycled over all facetting variables, (2) an unnamed character vector that must be of the same length as there are facetting variables (rows first, columns second), or (3) a named character vector where names refer to the faccetting variables to apply each character string.}

\item{parsed}{Optional names of the facetting variables that should have their labels parsed. See \code{?bquote} and the \code{plotmath} syntax documentation for how the labels should be formatted if they are parsed into mathematical expressions.}

\item{header}{Optional names of facetting variables to which to prepend their own column names. Overwrites \code{prepend}, if any.}

\item{sep}{Separator to use when prepending a \code{header}.}
}
\value{
A facetted ggplot
}
\description{
Customized facetted plot
}
\details{
Flexible tool for making facetted plots in various layouts and with various facet labels.
}
