% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facettize_check.R
\name{facettize_check}
\alias{facettize_check}
\title{Check input}
\usage{
facettize_check(arg, facets, default = "")
}
\arguments{
\item{arg}{The argument}

\item{facets}{A character vector of facetting variable names}

\item{default}{The default value to assign a facet}
}
\value{
A character vector
}
\description{
Checks and if needed reformats some input arguments to the \code{facettize} function.
}
