% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_facetlabel.R
\name{fct_facetlabel}
\alias{fct_facetlabel}
\title{Format a factor for use in plot facet}
\usage{
fct_facetlabel(x, prepend = "", append = "", surround = "")
}
\arguments{
\item{x}{A vector (will be coerced to a factor)}

\item{prepend, append, surround}{Character strings to use to prepend, append and surround the factor labels}
}
\value{
A factor
}
\description{
Prepends, appends and surrounds the labels of a factor with specific character strings
}
