% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.R
\name{ggboxplot}
\alias{ggboxplot}
\title{Plot a boxplot}
\usage{
ggboxplot(
  data,
  variable,
  x = NULL,
  mapping = NULL,
  is_jitter = FALSE,
  jitter_pars = list(),
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Name of the variable to plot}

\item{x}{Optional name of a categorial variable for the horizontal axis}

\item{mapping}{A named list of arguments to \code{ggplot2::aes_string}, containing
additional mappings}

\item{is_jitter}{Whether to add jittered points}

\item{jitter_pars}{Parameters to be passed to \code{ggplot2::geom_jitter}}

\item{...}{Parameters to be passed to \code{ggplot2::geom_boxplot}}
}
\value{
A ggplot

A ggplot
}
\description{
Plot a boxplot
}
\examples{

mtcars$vs <- factor(mtcars$vs)
ggboxplot(mtcars, "mpg", is_jitter = TRUE, jitter_pars = list(width = 0.2))
ggboxplot(mtcars, "mpg", mapping = list(fill = "vs"), alpha = 0.5)

}
