% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensity.R
\name{ggdensity}
\alias{ggdensity}
\title{Plot a density distribution}
\usage{
ggdensity(data, variable, mapping = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Name of the variable to plot}

\item{mapping}{A named list of arguments to \code{ggplot2::aes_string}, containing
additional mappings}

\item{...}{Parameters to be passed to \code{ggplot2::geom_density}}
}
\value{
A ggplot

A ggplot
}
\description{
Plot a density distribution
}
\examples{

mtcars$vs <- factor(mtcars$vs)
ggdensity(mtcars, "mpg")
ggdensity(mtcars, "mpg", mapping = list(fill = "vs"), alpha = 0.5)

}
