% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensityplot.R
\name{ggdensityplot}
\alias{ggdensityplot}
\title{Flexible density plots}
\usage{
ggdensityplot(
  data,
  variable,
  grouping = NULL,
  plot_type = "histogram",
  bins = 30,
  jitter = FALSE,
  jitter_width = 0.2,
  color_groups = TRUE,
  alpha = 0.5,
  colors = NULL
)
}
\arguments{
\item{data}{A data frame}

\item{variable}{What variable to plot the density of}

\item{grouping}{Optional grouping variable}

\item{plot_type}{Either of "density", "histogram", "boxplot" or "violin"}

\item{bins}{Number of bins in case of histogram}

\item{jitter}{Whether to add jitter in case of boxplot or violin}

\item{jitter_width}{Width of the jitter}

\item{color_groups}{Whether to color by group}

\item{alpha}{Transparency}

\item{colors}{Optional set of colors for the different groups}
}
\value{
A ggplot
}
\description{
A flexible function to explore different ways of plotting densities from a data frame
}
