% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfacet.R
\name{ggfacet}
\alias{ggfacet}
\title{Flexible plot facetting}
\usage{
ggfacet(
  p,
  facet_rows = NULL,
  facet_cols = NULL,
  wrap = FALSE,
  prepend = NULL,
  append = NULL,
  parsed = NULL,
  header = NULL,
  sep = " = "
)
}
\arguments{
\item{p}{A ggplot}

\item{facet_rows}{Optional variables to facet by rows}

\item{facet_cols}{Optional variables to facet by columns}

\item{wrap}{If TRUE, \code{facet_wrap} is used instead of \code{facet_grid}, so columns and rows are irrelevant}

\item{prepend, append}{Optional strings to prepend or append to the labels. See \code{?ggfacet_make}}

\item{parsed}{Optional names of the facetting variables that should have their labels parsed.}

\item{header}{Optional names of facetting variables to which to prepend their own column names}

\item{sep}{Separator to use when prepending a header.}
}
\value{
A facetted ggplot
}
\description{
Customized facetted plot
}
\details{
Flexible tool for making facetted plots in various layouts and with various facet labels.
}
