% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfacet_check.R
\name{ggfacet_check}
\alias{ggfacet_check}
\title{Check input}
\usage{
ggfacet_check(arg, facets, default = "")
}
\arguments{
\item{arg}{The argument}

\item{facets}{A character vector of facetting variable names}

\item{default}{The default value to assign a facet}
}
\description{
This function checks the validity of an argument passed to \code{ggfacet_make} and reformats it if needed for the function to be able to use.
}
