% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfacet_labs.R
\name{ggfacet_labs}
\alias{ggfacet_labs}
\title{Make labels for a facetting variable}
\usage{
ggfacet_labs(labels, prepend = "", append = "", parsed = FALSE)
}
\arguments{
\item{labels}{A list of labels for a specific facetting variable}

\item{prepend, append}{Optional strings to prepend or append to the labels, respectively.}

\item{parsed}{Whether the labels should be parsed into \code{plotmath} expressions.}
}
\value{
A list of new labels.
}
\description{
This function is used by \code{ggfacet_make} to make facet labels for each facetting variable.
}
