% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfacet_make.R
\name{ggfacet_make}
\alias{ggfacet_make}
\title{Customized facet labels}
\usage{
ggfacet_make(facets, prepend = NULL, append = NULL, parsed = FALSE)
}
\arguments{
\item{facets}{A list of labels for each facetting variable. Typically the kind of list that is passed to the labeller function if \code{labeller} is specified in \code{ggplot}'s facetting functions. See e.g. \code{?facet_grid}.}

\item{prepend, append}{Optional character vectors specifying strings to prepend or append to the facet labels, respectively. If named, the names should refer to specific facetting variables to which to apply this concatenation. If unnamed, the number of values must match the number of facetting variables, or be one, in which case the single value is recycled for all facetting variables.}

\item{parsed}{Optional logical vector specifying for which facetting variables should the labels be parsed into \code{plotmath} expressions. Same rules apply as for the previous arguments.}
}
\value{
A list of facet labels for each facetting variable.
}
\description{
Returns a list of facet labels for each facetting variable. This function can be passed to the \code{labeller} argument of \code{ggplot}'s facetting functions.
}
