% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Plot a heatmap}
\usage{
ggheatmap(data, variable, x, y, reduce = NULL, how = last, keep = NULL)
}
\arguments{
\item{data}{A data frame}

\item{variable}{What variable to plot?}

\item{x, y}{What variables for the x- and y-axes?}

\item{reduce}{Optional list of extra shrinking steps to perform prior to summarizing the data over all tiles (as defined by combining facets, \code{x} and \code{y}). See \code{?shrink}.}

\item{how}{Function used to summarize the data within each tile. Defaults to \code{last}. Irrelevant if the data already has one value for each tile. If \code{reduce} is specified, this argument must be provided explicitly as a list with one summary function for each grouping level in \code{reduce} and a final summary function for each tile. See \code{?shrink}.}

\item{keep}{Optional extra grouping variables to keep e.g. for facetting}
}
\value{
A ggplot
}
\description{
Generates a heatmap from a data frame
}
