% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghistogram.R
\name{gghistogram}
\alias{gghistogram}
\title{Plot a histogram}
\usage{
gghistogram(data, variable, mapping = NULL, ...)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Name of the variable to plot}

\item{mapping}{A named list of arguments to \code{ggplot2::aes_string}, containing
additional mappings}

\item{...}{Parameters to be passed to \code{ggplot2::geom_histogram}}
}
\value{
A ggplot

A ggplot
}
\description{
Plot a histogram
}
\examples{

mtcars$vs <- factor(mtcars$vs)
gghistogram(mtcars, "mpg")
gghistogram(mtcars, "mpg", mapping = list(fill = "vs"), alpha = 0.5)

}
