% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglineplot.R
\name{gglineplot}
\alias{gglineplot}
\title{Plot multiple lines}
\usage{
gglineplot(data, x, y, line = NULL, alpha = 0.5, step = 0.1, color = NULL)
}
\arguments{
\item{data}{A data frame}

\item{x, y}{Variables to plot against each other}

\item{line}{Variable to group data points into separate lines. If NULL (defaults), a single line is drawn per facet.}

\item{alpha}{Baseline transparency of the lines.}

\item{step}{Range of the noise in transparency across lines. Giving the lines different transparency values allows to plot many lines together on the same plot. Keep this parameter small to make the differences in transparency unnoticeable.}

\item{color}{Color of the lines. Leave unspecified if you want to assign color as an aesthetics (do that outside the function).}
}
\value{
A ggplot
}
\description{
Flexible way to plot multiple lines from a data frame
}
