% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviolin.R
\name{ggviolin}
\alias{ggviolin}
\title{Plot a violin plot}
\usage{
ggviolin(
  data,
  variable,
  x = NULL,
  mapping = NULL,
  is_jitter = FALSE,
  jitter_pars = list(),
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Name of the variable to plot}

\item{x}{Optional name of a categorial variable for the horizontal axis}

\item{mapping}{A named list of arguments to \code{ggplot2::aes_string}, containing
additional mappings}

\item{is_jitter}{Whether to add jittered points}

\item{jitter_pars}{Parameters to be passed to \code{ggplot2::geom_jitter}}

\item{...}{Parameters to be passed to \code{ggplot2::geom_violin}}
}
\value{
A ggplot


}
\description{
Plot a violin plot
}
\examples{

mtcars$vs <- factor(mtcars$vs)
ggviolin(mtcars, "mpg", is_jitter = TRUE, jitter_pars = list(width = 0.2))
ggviolin(mtcars, "mpg", mapping = list(fill = "vs"), alpha = 0.5)

}
