% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Serial shrinking}
\usage{
shrink(data, variables, grouping, how = last)
}
\arguments{
\item{data}{A data frame}

\item{variables}{Variables to summarize}

\item{grouping}{A list of single, or vectors of, grouping variables to be used sequentially. The groups are summarized in the order they are provided.}

\item{how}{A list of summarizing functions to be used, respectively, on each of the grouped data frames defined in \code{groups}}
}
\value{
A shrunk data frame
}
\description{
Sequentially applies \code{group_by} and \code{summarize} to a data frame.
}
