% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_prepend.R
\name{str_prepend}
\alias{str_prepend}
\title{Append a string to another}
\usage{
str_prepend(x, string, sep = "")
}
\arguments{
\item{x}{A character vector}

\item{string}{A character string to prepend}

\item{sep}{Optional separator string}
}
\value{
A character vector
}
\description{
Append a string to another
}
