% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_surround.R
\name{str_surround}
\alias{str_surround}
\title{Surround a string with another}
\usage{
str_surround(x, string = "")
}
\arguments{
\item{x}{A character vector of strings to surround}

\item{string}{Another character}
}
\value{
A character vector
}
\description{
Surround a string with another
}
