% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facettize.R
\name{facettize}
\alias{facettize}
\title{Flexible plot facetting}
\usage{
facettize(
  p,
  facet_rows = NULL,
  facet_cols = NULL,
  facet_wrapped = FALSE,
  prepend = NULL,
  append = NULL,
  labelling = NULL
)
}
\arguments{
\item{p}{A ggplot}

\item{facet_rows}{Optional variables to facet by rows}

\item{facet_cols}{Optional variables to facet by columns}

\item{facet_wrapped}{If TRUE, \code{facet_wrap} is used instead of \code{facet_grid}, so columns and rows are irrelevant}

\item{prepend, append}{Optional list of strings to prepend or append to the facet labels. The list should be named after the factors to prepend to.}

\item{labelling}{Optional list of functions to pass to \code{labeller} (e.g. \code{facet_value}, \code{label_both} or \code{label_parsed}).}
}
\value{
A facetted ggplot
}
\description{
Customizable way to split a ggplot into multiple facets depending on facetting variables.
}
