% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_append.R
\name{str_append}
\alias{str_append}
\title{Append a string to another}
\usage{
str_append(x, string, sep = "")
}
\arguments{
\item{x}{A string}

\item{string}{A string to prepend}

\item{sep}{Optional separator string}
}
\description{
Append a string to another
}
